split up navigation service into separate partial classes
